requirements.txt
# 深度学习框架
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# 数据处理
datasets>=2.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 模型相关
sentencepiece>=0.1.99
protobuf>=3.20.0
tokenizers>=0.13.0

# 评估指标
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1

# 数学计算
sympy>=1.12
networkx>=3.1

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 工具包
tqdm>=4.65.0
requests>=2.31.0
huggingface-hub>=0.15.0
pyyaml>=6.0
argparse

# Jupyter环境
jupyter>=1.0.0
ipywidgets>=8.0.0

# 其他工具
psutil>=5.9.0
memory-profiler>=0.60.0

# OpenAI API
openai==0.28.0 